LayoutTests/fast/js/dfg-double-use-of-post-simplification-double-prediction-expected.txt

Tests stability of the DFG compiler when you have a double use of a variable that is not revealed to be a double until after CFG simplification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS foo(0.5) is 42.5
PASS successfullyParsed is true

TEST COMPLETE
